Allow Balancer SOR solver on Gnosis Chain (#1779)
It seems that the old Balancer SOR API is actually working on Gnosis
Chain now. This would help us with a bunch of headaches around
Karpatekey's liquidity provision needs as the SOR solver supports all
Balancer pools.

### Test Plan

Run with orderbook on GChain with just Baseline and BalancerSor as
estimators. See that selling WETH for wxDAI works (with Balancer being
the winning solver).

### Release notes

Once merged I plan to enable Balancer as a solver (and then price
estimator) on staging