Move ethrpc into separate crate (#1804)
Moves `ethrpc` into its own crate because it's self contained and this
way we can easily reuse it in our different crates without pulling it
from `shared`.
Over all the idea is to trim down `shared`.

This PR tries to move the code with minimal changes. Whether or not code
inside the new crate should be refactored or removed should be evaluated
in a separate PR.

### Test Plan
No behavior changes so existing CI and e2e tests should be enough.