Account for unstable order of `select_all` (#1882)
The docs for `select_all` say:
> There are no guarantees provided on the order of the list with the
remaining futures. They might be swapped around, reversed, or completely
random.

When simplifying the price API I was not aware of this and didn't take
any precautions to account for any unstable order of the returned
futures.

I checked the implementation of `select_all` and it uses
`Vec::swap_remove()` under the hood which definitely messes with the
order of the returned futures and the indices.
It is possible that this is the reason why we saw weird 95th percentile
quoting latency. The new code returns the index along side the quote
result and does not use the index returned by `select_all` at all.