Native  price estimator using 1Inch spot token API (#1876)
Creates a native price estimator using the [1Inch spot token
API](https://portal.1inch.dev/documentation/spot-price/swagger). It is
querying the list of tokens periodically in the background (getting
~1300 per request) and allows fast lookup for this list via a HashMap.
The API even allows to query for additional tokens that are not part of
this list (e.g.
[HarryPotterObamaSonic10Inu](https://etherscan.io/token/0x72e4f9f808c49a2a61de9c5896298920dc4eeea9))
so we can even potentially improve the hit rate by recording requests
that weren't in the default list.

The idea for this estimator is to work as a cheap first line of defence
(together with the baseline estimator and potentially Quasimodo, cf
#1841) to avoid more expensive API requests on native token queries.

It's not yet hooked up, because I'm still wondering if we should use the
general `PriceEstimating` trait (which requires additional sanity checks
that we are indeed receiving a native price estimate) or the more
specific `NativePriceEstimating` trait.
Currently we do not have a `RacingCompetitionPriceEstimator` component
for `NativePriceEstimating` (only `PriceEstimating`) and I'm not sure
what the best way is to implement it without repeating this non trivial
component. Technically we could use the spot price also to compute
generic "estimates" (but they will likely not be accurate) and the
interface currently requires shoehorning a zero gas amount and unknown
solver which is kind of ugly.

Looking for suggestions!

### Test Plan

Create you own developer account + API Key, then 
```
export ONEINCH_AUTH_TOKEN=...
cargo test --package shared --lib -- price_estimation::native::oneinch::tests::works --exact --nocapture --ignored
```

---------

Co-authored-by: Dusan Stanivukovic <dusan.stanivukovic@gmail.com>