Scale Orders By Remaining Amount And Available Balance (#1902)
# Description

This PR essentially combines #1861 and #1900 into a single PR. The
reason for providing an alternative PR is that some of the decisions
taking in #1900 (on which #1861 is based on) are directly related to the
requirements of #1861. I hope this makes reviewing easier 🤞.

This PR fixes an issue where partially fillable order amounts were not
being scaled before passing to the solver engines as expected, leading
to invalid settlements.

# Changes

1. The order `Partial` now carries an available amount and not how much
was executed. This allows us to encode things like the remaining order
balance in the same field (so "available" here means how much of a
partially fillable order is left and covered by user balance).
2. Use the scaled available amounts to pass to the solvers instead of
the full amounts.

## How to test

Rust CI - added unit tests to verify available amount scaling logic. An
E2E test that fully checks the balance allocation is under development
and will be proposed in a follow up PR.