Simplify While Loop in Autopilot RunLoop (#1933)
# Description

Follow up to
https://github.com/cowprotocol/services/pull/1930#discussion_r1349967216.

This PR removes the extra `while` loop logic in `autopilot::RunLoop`, as
it was unneeded and confusing.

The issue that the comment that existed alluded to was that when you do
something like:

```rust
while let _ = thing() {
    // body
}
```

```rust
match thing() {
    _ => {
        // body
    }
}
```

`thing()` is not dropped until the end of the `while` or `match` (even
if there is no binding to it in the matched pattern). This results in
`Drop::drop` only getting called at the end of the `while`/`match` which
may be unexpected and lead to deadlocks.

This is **not** the case here for a couple of reasons:
- We are using a while expression (`while some_bool_express()`) and not
while with pattern matching (`while let ... = some_expression()`)
- The `borrow()` is only used as a temporary value as part of the
boolean expression and gets dropped immediately.

We can write some [Rust
code](https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=ff32cf291c63459b065f9d99488baf90)
to prove this:

```rust
mod foo {
    pub struct Foo {
        pub value: u64,
    }

    impl Foo {
        pub fn borrow(&self) -> Bar<'_> {
            Bar(self)
        }
    }

    pub struct Bar<'a>(&'a Foo);

    impl std::ops::Deref for Bar<'_> {
        type Target = Foo;
        fn deref(&self) -> &Self::Target {
            self.0
        }
    }

    impl Drop for Bar<'_> {
        fn drop(&mut self) {
            println!("Bar::drop");
        }
    }
}

fn main() {
    let mut foo = foo::Foo { value: 0 };
    println!("main enter");
    while let 42 = foo.borrow().value {
        println!("main inner");
        foo.value = 42;
    }
    println!("main exit");
}
```

Which outputs:

```
main enter
Bar::drop
main inner
Bar::drop
main exit
```

Meaning that `Bar` does indeed get dropped after every `while`
expression computation.

Note that if we were to change the above example code on line 31 to:

```rust
    while let 0 = foo.borrow().value {
```

We would get a compiler error for:

```
error[E0506]: cannot assign to `foo.value` because it is borrowed
  --> src/main.rs:33:9
   |
31 |     while let 0 = foo.borrow().value {
   |                   ------------
   |                   |
   |                   `foo.value` is borrowed here
   |                   a temporary with access to the borrow is created here ...
32 |         println!("main inner");
33 |         foo.value = 42;
   |         ^^^^^^^^^^^^^^ `foo.value` is assigned to here but it was already borrowed
34 |     }
   |     - ... and the borrow might be used here, when that temporary is dropped and runs the `Drop` code for type `Bar`
```

Which corroborates that this issue only applies to `while let` pattern
matching and not `while` with boolean expressions.

# Changes

- [x] Remove unnecessary manual `drop` call in `while` loop expression.

## How to test

See Rust playground link to verify these are the expected Rust
semantics.