Allow U256 values to be deserialized either from hex or decimal (#1913)
# Description
We were made aware that the interaction's value field is parsed using
the default U256 string parser (which assumes fields to be in base16
represenation). This was an oversight on our end and is being fixed in
this PR by interpreting the value as
- base16 string if prefixed by 0x
- base10 string otherwise

We also use the new type on most other "amount" fields in the solver
API.

Assuming that no solvers are currently using the value field with non
prefixed hex strings, this change should be backwards compatible. It
makes previous `u256_dec` deserialized values more permissive (allowing
to parse them from prefixed hex strings as well) while keeping their
serialisation behavior.

# Changes
- Implemented u256_hex_or_decimal deserialization (serialization same as
u256_dec)
- Use the new type for most value fields on the solver response

## How to test
Added a unit test for the serialization

## Related Issues

Fixes #1910