Add metadata.signer field to app data and reject mismatching orders (#1939)
# Description

Adds a new field to the app data, `metadata.signer`, that stores the
address for which the order is signed.
If `metadata.signer` doesn't match the owner of the order, the order is
rejected at creation time.
Note that it's still possible to omit the signer field or not provide
some explicit metadata, in which case the order is still accepted.

Note that the changes also apply to order replacement, since it follows
the same code path as order creation; this is however untested.

The new field follows the exact same logic as the `from` field in the
order creation. We could consider deprecating the `from` field once we
make providing app data mandatory.

[Related
document](https://docs.google.com/document/d/1OpWXRZyqaQjEQqzX_d4AI5iphRV0Z4k5fCHnEs5jWwo/edit#heading=h.8aw19fkz5xgi).

# Next steps

Right now, it's possible to create an order without explicit app data
that is later found to be from an invalid owner. A future PR will add an
extra state to the order that marks it as invalid if an order becomes
invalid when new information on the app data becomes available.

## How to test

New e2e test.