Correctly exclude paraswap dexs (#1990)
# Description
The paraswap solver issues a lot of
[warnings](https://production-6de61f.kb.eu-central-1.aws.cloud.es.io/app/discover#/?_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:'2023-10-19T09:19:08.195Z',to:'2023-10-19T09:28:29.622Z'))&_a=(columns:!(log),filters:!(('$state':(store:appState),meta:(alias:!n,disabled:!f,index:ea511870-d9b3-11ed-a9d0-a17451f01cc1,key:kubernetes.container_name,negate:!f,params:(query:paraswap),type:phrase),query:(match_phrase:(kubernetes.container_name:paraswap)))),grid:(columns:('@timestamp':(width:164))),index:ea511870-d9b3-11ed-a9d0-a17451f01cc1,interval:auto,query:(language:kuery,query:'log:%20%22It%20is%20not%20allowed%20to%20have%20limit-orders%20Dex%20or%20paraswappool%20Dex%20in%20route%22'),sort:!(!('@timestamp',desc)))):
```
It is not allowed to have limit-orders Dex or paraswappool Dex in route and
'ignoreChecks=true' together. Consider requesting the price with excluded
limit-orders if you want to use 'ignoreChecks': &excludeDEXS=ParaSwapPool,ParaSwapLimitOrders
```
We already had the machinery required to append those excluded dexs as a
query parameter but we didn't get the capitalization right which lead to
the argument being ignored by the paraswap API.

One thing to note is that only the `/prices` API supports disabling dexs
but not the `/transactions` API.

# Changes
* changes the serialized query parameter `excludeDexs` to `excludeDEXS`.
* also mentions which 2 dexs always have to be excluded in the example
config file

## How to test
Run paraswap in colocated mode to see that the warnings stop.