Make /settle call blocking and report tx_hash (#1999)
# Description
Currently the autopilot wastes a lot of time waiting for transactions
that will never appear. This happens because the driver's `/settle`
endpoint operates in a fire-and-forget (get request, kick off submission
in background, return immediately). That way the driver does not have
any way to communicate that it will not be able to submit the solution
to the autopilot and it will have to monitor the blockchain until the
deadline is reached.

This PR is slightly related to
https://github.com/cowprotocol/services/issues/1974 but does not solve
the issue that solutions don't get submitted in the first place.

# Changes
This PR makes it so that `/settle` actually blocks in the driver until
the solution gets submitted and returns the tx_hash. If the driver is
not able to submit the solution (e.g. the simulations for the solution
start reverting) it will return an error to the autopilot which will
immediately move to the next auction.

Note that this is not really how it's supposed to work but this solution
is fine as long as we are running all the drivers (and can therefore
assume a reasonable behavior). We should revisit this decision before we
can encourage external parties to run their own driver.

## How to test
e2e tests